refactor: 예외 패키지 정리 (#54)
